Fix date problems

+++ backend/app/repositories/models/enterprise.py
@@ -27,7 +27,7 @@ class EnterpriseModel(BaseModel):
     used_licenses: int = Field(default=0, ge=0, description="Currently used licenses")
     contract_start_date: Optional[datetime] = Field(None, description="Contract start date (YYYY-MM-DD)")
     contract_end_date: Optional[datetime] = Field(None, description="Contract end date (YYYY-MM-DD)")
-    monthly_revenue: Optional[float] = Field(default=0, ge=0, description="Monthly revenue from this enterprise")
+    monthly_revenue: Optional[int] = Field(default=0, ge=0, description="Monthly revenue from this enterprise")
 
     # Added
     created_date: datetime = Field(..., description="Creation date (YYYY-MM-DD)")
