+++ backend/app/routes/schemas/entreprise.py
@@ -53,7 +53,7 @@ class EnterpriseInput(BaseSchema):
     used_licenses: int = Field(default=0, ge=0, description="Currently used licenses")
     contract_start_date: datetime = Field(..., description="Contract start date (YYYY-MM-DD)")
     contract_end_date: Optional[datetime] = Field(None, description="Contract end date (YYYY-MM-DD)")
-    monthly_revenue: Optional[float] = Field(default=0, ge=0, description="Monthly revenue from this enterprise")
+    monthly_revenue: Optional[int] = Field(default=0, ge=0, description="Monthly revenue from this enterprise")
     
     @validator('website')
     def validate_website(cls, v):
@@ -90,7 +90,7 @@ class EnterpriseUpdate(BaseSchema):
     used_licenses: Optional[int] = Field(None, ge=0, description="Currently used licenses")
     contract_start_date: Optional[datetime] = Field(None, description="Contract start date (YYYY-MM-DD)")
     contract_end_date: Optional[datetime] = Field(None, description="Contract end date (YYYY-MM-DD)")
-    monthly_revenue: Optional[float] = Field(None, ge=0, description="Monthly revenue from this enterprise")
+    monthly_revenue: Optional[int] = Field(None, ge=0, description="Monthly revenue from this enterprise")
     
     @validator('website')
     def validate_website(cls, v):
@@ -122,7 +122,7 @@ class EnterpriseOutput(BaseSchema):
     used_licenses: int = Field(..., description="Currently used licenses")
     contract_start_date: datetime = Field(None, description="Contract start date (YYYY-MM-DD)")
     contract_end_date: Optional[datetime] = Field(None, description="Contract end date (YYYY-MM-DD)")
-    monthly_revenue: Optional[float] = Field(None, description="Monthly revenue from this enterprise") 
+    monthly_revenue: Optional[int] = Field(None, description="Monthly revenue from this enterprise") 
 
     # added
     created_date: datetime = Field(..., description="Creation date (YYYY-MM-DD)")
